requirements.txt
camelot-py[cv]
markitdown[all]
mcp[cli]
toml